Remove cast for noise dtype

+++ vlm/train/MM-Grounding-DINO/mmdetection/mmdet/datasets/transforms/gaussian_noise.py
@@ -35,6 +35,6 @@ class GaussianNoise(ColorTransform):
         img = results['img']
         img_dtype = results['img'].dtype
         
-        img = img.astype(np.float32) + np.random.normal(0., mag, img.shape, dtype=np.float32)
+        img = img.astype(np.float32) + np.random.normal(0., mag, img.shape)
         results['img'] = np.clip(img, 0., 255.).astype(img_dtype)
